Move get_metric to evaluate module
